Extract wizard runtime into top-level module
