--- gkc/profiles/forms/wizard/__init__.py
@@ -1,13 +1,0 @@
-"""Wizard step implementations.
-
-Plain meaning: The individual screens/pages of the multi-step form.
-"""
-
-from gkc.profiles.forms.wizard.step_base import Step
-from gkc.profiles.forms.wizard.steps import (
-    IdentificationStep,
-    SitelinksStep,
-    StatementsStep,
-)
-
-__all__ = ["Step", "IdentificationStep", "SitelinksStep", "StatementsStep"]
