+++ gkc/wizard/draft_manager.py
